requirements.txt
Flask
Pillow
numpy
tensorflow==2.15.0
requests
